Code cleanup

+++ src/main/java/comeon/ui/pictures/ConstrainedAxis.java
@@ -7,15 +7,15 @@ enum ConstrainedAxis {
   VERTICAL {
     @Override
     Dimension getPreviewPanelDimension(final BufferedImage image, final int desiredSize) {
-      return new Dimension((int) ((double) image.getWidth() * ((double) desiredSize / (double) image.getHeight())), desiredSize);
+      return new Dimension((int) (image.getWidth() * ((double) desiredSize / (double) image.getHeight())), desiredSize);
     }
   },
   HORIZONTAL {
     @Override
     Dimension getPreviewPanelDimension(final BufferedImage image, final int desiredSize) {
-      return new Dimension(desiredSize, (int) ((double) image.getHeight() * ((double) desiredSize / (double) image.getWidth())));
+      return new Dimension(desiredSize, (int) (image.getHeight() * ((double) desiredSize / (double) image.getWidth())));
     }
   };
-  
+
   abstract Dimension getPreviewPanelDimension(BufferedImage image, int desiredSize);
 }
